+++ src/server/rest/v1/service/TenantService.ts
@@ -1,7 +1,7 @@
 import { Action, Entity } from '../../../../types/Authorization';
+import { CryptoSettingsType, KeySetting, SettingDB, SettingDBContent } from '../../../../types/Setting';
 import { HTTPAuthError, HTTPError } from '../../../../types/HTTPError';
 import { NextFunction, Request, Response } from 'express';
-import { SettingDB, SettingDBContent } from '../../../../types/Setting';
 import Tenant, { TenantLogo } from '../../../../types/Tenant';
 import User, { UserRole } from '../../../../types/User';
 
@@ -17,6 +17,7 @@ import { ServerAction } from '../../../../types/Server';
 import SettingStorage from '../../../../storage/mongodb/SettingStorage';
 import SiteAreaStorage from '../../../../storage/mongodb/SiteAreaStorage';
 import { StatusCodes } from 'http-status-codes';
+import TenantComponents from '../../../../types/TenantComponents';
 import TenantStorage from '../../../../storage/mongodb/TenantStorage';
 import TenantValidator from '../validator/TenantValidation';
 import UserStorage from '../../../../storage/mongodb/UserStorage';
@@ -163,6 +164,28 @@ export default class TenantService {
     next();
   }
 
+  public static async createInitialSettingsForTenant(tenantID: string): Promise<void> {
+    await this.createInitialCryptoSetting(tenantID);
+  }
+
+  public static async createInitialCryptoSetting(tenantID: string): Promise<void> {
+    // Check for settings in db
+    const keySettings = await SettingStorage.getCryptoSettings(tenantID);
+    // Generate Crypto Key Settings
+    if (!keySettings) {
+      const keySettingToSave = {
+        identifier: TenantComponents.CRYPTO,
+        type: CryptoSettingsType.CRYPTO,
+        crypto: {
+          key: Utils.generateKey(),
+          keyProperties: Utils.getDefaultKeyProperties()
+        }
+      } as KeySetting;
+      // Save Crypto Key Settings
+      await SettingStorage.saveCryptoSettings(tenantID, keySettingToSave);
+    }
+  }
+
   public static async handleCreateTenant(action: ServerAction, req: Request, res: Response, next: NextFunction): Promise<void> {
     // Validate
     const filteredRequest = TenantValidator.getInstance().validateTenantCreateRequestSuperAdmin(req.body);
@@ -212,6 +235,8 @@ export default class TenantService {
     if (await LockingManager.acquire(createDatabaseLock)) {
       try {
         await TenantStorage.createTenantDB(filteredRequest.id);
+        // Create initial settings for tenant
+        await TenantService.createInitialSettingsForTenant(filteredRequest.id);
       } finally {
         // Release the database creation Lock
         await LockingManager.release(createDatabaseLock);
